uclibc: updates for not (yet) used features
This commit bundles upstream changes:

* for features not used by L4Re,
* implemented differently on L4Re
* for architectures not currently supported by L4Re
* introducing new features / functions to uclibc-ng which are not used
  by L4Re (yet)

It includes the following commits:

--

posix_fadvise: handle 2 variants for SYSCALL_ALIGN_64BIT
sync_file_range: fix standard UCLIBC_SYSCALL_ALIGN_64BIT handling

Cc: Baruch Siach <baruch@tkos.co.il>
Cc: Max Filippov <jcmvbkbc@gmail.com>

--

sparc/clone.S: guard tcb-offsets.h include with RESET_PID


--

libc: update strverscmp
libc: add getrandom(2)
SH: add 't' to syscall clobber list
NPTL: SH: silence warning
libc: Fix page-size in getifaddrs()
librt: Add missing __dso_handle
nptl_db: fix ommitting td_ta_setconcurrency
getconf.c: undef VERSION


--

nptl/atfork: Hide pthread_atfork in shared versions


--

init.c (__nptl_initial_report_events): New variable.


--

Fix libgcc_s_resume issue with gcc 5.1


--

nptl_db/db_info: fix the incorrect initial size for dtvp


--

libc: fix setting return value and errno in fallocate()
libc: fix sign extension in fallocate()
libc: posix_fallocate must return an error number on failure


--

fanotify: add system call support


--

Do not support localised optimizations this helps to have a global -O level


--

svc.c: svc_getreqset() buffer overflow
dynamically allocate ahostbuf buffer


--

Fix syslog messages lost after syslogd restarts
Fix syslog messages lost if syslogd is temporary busy


--

Add implementation for copysignl for ppc


--

use arch specific thread stack size for nptl
Fix "pselect: Use linux pselect6 syscall when available"
Add implementation for copysignl for i386
NPTL: Fix __pthread_once old style declaration
NPTL: fix infinite recursion

Reported-by: Leonid Lisovskiy <lly.dev@gmail.com>

--

pselect: Use linux pselect syscall when available


--

openat argument fix.
Remove alias.


--

use weak to fix f.e. cdrkit static compile. fixes #3
fix static linking of pthread apps
just use 4k pages for microblaze, the config are missing and never committed into uClibc
integrate old m68k vfork bugfix of pre-µClibc-ng tree
fix inline asm changing the stack pointer leading to segfaults problem
sh: fix static linking issue
locales: add patch from OpenWrt
add support for euidaccess/eaccess legacy functions
fanotify: include needed Linux definitions
i386: use socketcall even if newer linux exposes direct syscalls
bfin: use vfork.S again
bfin: fix clashing with newer gcc
metag: add missing header file
sh: add back sh2 specific assembly for clone()
libc/stdlib: canonicalize_file_name() memory leak
reduce stack cache size to save memory space
fix bfin compile, -msoft-float is not available

--

+++ uclibc/lib/contrib/uclibc/libpthread/nptl/sysdeps/unix/sysv/linux/sh/lowlevellock.h
@@ -358,7 +358,7 @@ extern int __lll_unlock_wake (int *__futex, int private) attribute_hidden;
     register unsigned long __r5 __asm__ ("r5")				      \
       = __lll_private_flag (FUTEX_WAIT, private);			      \
     register unsigned long __r6 __asm__ ("r6") = (unsigned long) (val);	      \
-    register unsigned long __r7 __asm__ ("r7") = (timeout);			      \
+    register unsigned long __r7 __asm__ ("r7") = (unsigned long) (timeout);   \
     __asm__ __volatile__ (SYSCALL_WITH_INST_PAD				      \
 		      : "=z" (__status)					      \
 		      : "r" (__r3), "r" (__r4), "r" (__r5),		      \
